+++ .eslintrc.json
@@ -10,7 +10,8 @@
     "react/jsx-filename-extension": ["error", { "extensions": [".tsx"] }],
     "no-unused-vars": "off",
     "no-undef": "off",
-    "import/named": "off"
+    "import/named": "off",
+    "jsx-a11y/label-has-for": "off"
   },
   "parserOptions": {
     "ecmaVersion": 2018,
