+++ src/components/atoms/Forms/TextInput.tsx
@@ -7,27 +7,27 @@ import SvgIcon from 'components/atoms/Icons/SvgIcon';
 
 import { color } from 'constants/styles';
 
-interface Props {
+type TextInputType = 'text' | 'password' | 'email';
+
+export interface TextInputProps {
   value?: string;
   name?: string;
-  label?: string;
   placeholder?: string;
   className?: string;
   warning?: boolean;
-  type?: string;
+  type?: TextInputType;
   onChange?(value: string): void;
 }
 
 export default ({
   value,
   name,
-  label,
   placeholder,
   className,
   onChange,
   warning,
   type,
-}: Props) => {
+}: TextInputProps) => {
   const [currentValue, setCurrentValue] = React.useState('');
   const [inFocus, setInFocus] = React.useState(false);
 
@@ -55,61 +55,50 @@ export default ({
   }, [value]);
 
   return (
-    <div className={className}>
-      <label htmlFor={name}>
-        <div
-          css={css`
-            font-size: 1.2rem;
-            padding: 1rem 0;
-          `}
-        >
-          {label}
-        </div>
-        <div
-          css={css`
-            position: relative;
-          `}
-        >
-          <input
-            value={currentValue}
-            css={css`
-              border-radius: 1000px;
-              border: solid 2px ${color.BLACK};
-              padding: 1rem 3rem 1rem 1.5rem;
-              font-size: 1.4rem;
-              outline: 0;
-              width: 100%;
-              color: inherit;
+    <div
+      className={className}
+      css={css`
+        position: relative;
+      `}
+    >
+      <input
+        value={currentValue}
+        css={css`
+          border-radius: 1000px;
+          border: solid 2px ${color.BLACK};
+          padding: 1rem 3rem 1rem 1.5rem;
+          font-size: 1.4rem;
+          outline: 0;
+          width: 100%;
+          color: inherit;
 
-              ::placeholder {
-                font-size: 1rem;
-                transform: translate(0, -0.1rem);
-              }
-            `}
-            onFocus={handleFocus}
-            onBlur={handleBlur}
-            placeholder={inFocus ? '' : placeholder}
-            name={name}
-            type={type || 'text'}
-            onChange={handleChange}
-          />
-          {warning && (
-            <SvgIcon
-              css={css`
-                position: absolute;
-                width: 1.5rem;
-                height: 1.5rem;
-                top: 0;
-                right: 1.2rem;
-                bottom: 0;
-                margin: auto;
-              `}
-              icon="warning"
-              alt="エラー"
-            />
-          )}
-        </div>
-      </label>
+          ::placeholder {
+            font-size: 1rem;
+            transform: translate(0, -0.1rem);
+          }
+        `}
+        onFocus={handleFocus}
+        onBlur={handleBlur}
+        placeholder={inFocus ? '' : placeholder}
+        id={name}
+        type={type || 'text'}
+        onChange={handleChange}
+      />
+      {warning && (
+        <SvgIcon
+          css={css`
+            position: absolute;
+            width: 1.5rem;
+            height: 1.5rem;
+            top: 0;
+            right: 1.2rem;
+            bottom: 0;
+            margin: auto;
+          `}
+          icon="warning"
+          alt="エラー"
+        />
+      )}
     </div>
   );
 };
